+++ src/components/Header.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+
+const Header = () => {
+  const headerStyle = {
+    padding: '20px 0',
+    lineHeight: '1.5em',
+  };
+
+  return (
+    <header style={headerStyle}>
+      <h1 className="title">
+        todos
+      </h1>
+    </header>
+  );
+};
+
+export default Header;
